move common type
