+++ packages/core/src/database/postgres/service.ts
@@ -1,4 +1,3 @@
-import type { Common } from "@/Ponder.js";
 import {
   type FunctionIds,
   HASH_VERSION,
@@ -9,6 +8,7 @@ import {
   getTableAccessInverse,
   isWriteStoreMethod,
 } from "@/build/static/getTableAccess.js";
+import type { Common } from "@/common/common.js";
 import { NonRetryableError } from "@/common/errors.js";
 import type { Schema } from "@/schema/types.js";
 import { isEnumColumn, isManyColumn, isOneColumn } from "@/schema/utils.js";
