+++ packages/core/src/database/sqlite/service.ts
@@ -1,5 +1,4 @@
 import path from "node:path";
-import type { Common } from "@/Ponder.js";
 import {
   type FunctionIds,
   HASH_VERSION,
@@ -10,6 +9,7 @@ import {
   getTableAccessInverse,
   isWriteStoreMethod,
 } from "@/build/static/getTableAccess.js";
+import type { Common } from "@/common/common.js";
 import { NonRetryableError } from "@/common/errors.js";
 import type { Schema } from "@/schema/types.js";
 import { isEnumColumn, isManyColumn, isOneColumn } from "@/schema/utils.js";
